+++ kippo/commands/__init__.py
@@ -18,4 +18,6 @@ __all__ = [
     'malware',
     'netstat',
     'which',
+    'gcc',
+    'iptables'
     ]