+++ CHANGELOG.md
@@ -24,7 +24,6 @@
 * fixes for behavior with non-existent files (cd /test, cat /test/nonexistent, etc)
 * fix for ability to ping/ssh non-existent IP address
 * always send ssh exit-status 0 on exec and shell
-* add 1st version of 'netstat'
 * ls output is now alphabetically sorted
 * banner_file is deprecated. honeyfs/etc/issue.net is default
 * add 'dir' alias for 'ls'
@@ -32,4 +31,5 @@
 * add 'users' aliased to 'whoami'
 * add 'killall' and 'killall5' aliased to nop
 * add 'poweroff' 'halt' and 'reboot' aliases for shutdown
-* add 'which' and environment passing to commands
+* add environment passing to commands
+* added 'which', 'netstat' and 'gcc' from kippo-extra